move logic to profile ops
