+++ core/src/tasks/profile/confirm.ts
@@ -1,16 +1,10 @@
-import { Op } from "sequelize";
 import Moment from "moment";
 import { CLSTask } from "../../classes/tasks/clsTask";
 import { CLS } from "../../modules/cls";
-import { Property } from "../../models/Property";
 import { plugin } from "../../modules/plugin";
-import { ProfileProperty } from "../../models/ProfileProperty";
-import { Profile } from "../../models/Profile";
 import { Schedule } from "../../models/Schedule";
 import { Run } from "../../models/Run";
 import { ProfileOps } from "../../modules/ops/profile";
-import { GroupMember } from "../../models/GroupMember";
-import { Import } from "../../models/Import";
 
 export class ProfilesConfirm extends CLSTask {
   constructor() {
@@ -20,62 +14,6 @@ export class ProfilesConfirm extends CLSTask {
     this.frequency = 1000 * 30;
     this.queue = "profiles";
     this.inputs = {};
-  }
-
-  async confirmProfileProperties(
-    limit: number,
-    fromDate: Date,
-    sourceId?: string
-  ) {
-    const properties = await Property.findAllWithCache();
-    const directlyMapped = properties.filter(
-      (p) => p.directlyMapped && (!sourceId || sourceId === p.sourceId)
-    );
-
-    const profiles = await Profile.findAll({
-      limit,
-      where: { state: "ready" },
-      include: [
-        GroupMember,
-        {
-          model: ProfileProperty,
-          required: true,
-          where: {
-            state: "ready",
-            confirmedAt: {
-              [Op.lt]: fromDate,
-            },
-            rawValue: {
-              [Op.ne]: null,
-            },
-            propertyId: directlyMapped.map((p) => p.id),
-          },
-        },
-      ],
-    });
-
-    const now = new Date();
-
-    const bulkImports = [];
-    for (const profile of profiles) {
-      delete profile.profileProperties; // get all profile properties
-      const oldProfileProperties = await profile.simplifiedProperties();
-      const oldGroupIds = profile.groupMembers.map((gm) => gm.groupId);
-
-      bulkImports.push({
-        profileId: profile.id,
-        profileAssociatedAt: now,
-        oldProfileProperties,
-        oldGroupIds,
-        creatorType: "task",
-        creatorId: this.name,
-      });
-    }
-
-    await Import.bulkCreate(bulkImports);
-    await ProfileOps.markPendingByIds(profiles.map((p) => p.id));
-
-    return profiles.length;
   }
 
   async runWithinTransaction() {
@@ -105,16 +43,16 @@ export class ProfilesConfirm extends CLSTask {
 
     for (const schedule of schedules) {
       const latestRunAt = schedule.runs[0].completedAt;
-      count += await this.confirmProfileProperties(
+      count += await ProfileOps.confirmExistence(
         limit - count,
         latestRunAt,
         schedule.sourceId
       );
     }
 
-    const days = 1;
+    const days = 1; // TODO make a setting
     const nextConfirmAt = Moment().subtract(days, "days").toDate();
-    count += await this.confirmProfileProperties(limit - count, nextConfirmAt);
+    count += await ProfileOps.confirmExistence(limit - count, nextConfirmAt);
 
     // re-enqueue if there is more to do
     if (count > 0) await CLS.enqueueTask(this.name, {});
